+++ example-src/cljs/example/main.cljs
@@ -6,13 +6,10 @@
             example.autocomplete-tree
             example.calendar
             example.timepicker
-            example.currency-input
             example.search-highlight
             example.dropdown
             example.input
-            example.form-group
             example.modal
-            example.nav
             example.autosize
             example.timeago
             example.clipboard